+++ packages/ai/core/registry/index.ts
@@ -1,3 +1,7 @@
 export { customProvider, experimental_customProvider } from './custom-provider';
 export { NoSuchProviderError } from './no-such-provider-error';
-export { experimental_createProviderRegistry } from './provider-registry';
+export {
+  createProviderRegistry,
+  experimental_createProviderRegistry,
+} from './provider-registry';
+export type { ProviderRegistryProvider } from './provider-registry';