+++ packages/ai/core/registry/provider-registry.ts
@@ -1,37 +1,102 @@
 import { NoSuchModelError, ProviderV1 } from '@ai-sdk/provider';
-import { EmbeddingModel, ImageModel, LanguageModel, Provider } from '../types';
+import { EmbeddingModel, ImageModel, LanguageModel } from '../types';
 import { NoSuchProviderError } from './no-such-provider-error';
+
+type ExtractLiteralUnion<T> = T extends string
+  ? string extends T
+    ? never
+    : T
+  : never;
+
+export interface ProviderRegistryProvider<
+  PROVIDERS extends Record<string, ProviderV1> = Record<string, ProviderV1>,
+  SEPARATOR extends string = ':',
+> {
+  languageModel<KEY extends keyof PROVIDERS>(
+    id: KEY extends string
+      ? `${KEY & string}${SEPARATOR}${ExtractLiteralUnion<Parameters<NonNullable<PROVIDERS[KEY]['languageModel']>>[0]>}`
+      : never,
+  ): LanguageModel;
+  languageModel<KEY extends keyof PROVIDERS>(
+    id: KEY extends string ? `${KEY & string}${SEPARATOR}${string}` : never,
+  ): LanguageModel;
+
+  textEmbeddingModel<KEY extends keyof PROVIDERS>(
+    id: KEY extends string
+      ? `${KEY & string}${SEPARATOR}${ExtractLiteralUnion<Parameters<NonNullable<PROVIDERS[KEY]['textEmbeddingModel']>>[0]>}`
+      : never,
+  ): EmbeddingModel<string>;
+  textEmbeddingModel<KEY extends keyof PROVIDERS>(
+    id: KEY extends string ? `${KEY & string}${SEPARATOR}${string}` : never,
+  ): EmbeddingModel<string>;
+
+  imageModel<KEY extends keyof PROVIDERS>(
+    id: KEY extends string
+      ? `${KEY & string}${SEPARATOR}${ExtractLiteralUnion<Parameters<NonNullable<PROVIDERS[KEY]['imageModel']>>[0]>}`
+      : never,
+  ): ImageModel;
+  imageModel<KEY extends keyof PROVIDERS>(
+    id: KEY extends string ? `${KEY & string}${SEPARATOR}${string}` : never,
+  ): ImageModel;
+}
 
 /**
  * Creates a registry for the given providers.
  */
-export function experimental_createProviderRegistry(
-  providers: Record<string, ProviderV1>,
-): Provider {
-  const registry = new DefaultProviderRegistry();
+export function createProviderRegistry<
+  PROVIDERS extends Record<string, ProviderV1>,
+  SEPARATOR extends string = ':',
+>(
+  providers: PROVIDERS,
+  {
+    separator = ':' as SEPARATOR,
+  }: {
+    separator?: SEPARATOR;
+  } = {},
+): ProviderRegistryProvider<PROVIDERS, SEPARATOR> {
+  const registry = new DefaultProviderRegistry<PROVIDERS, SEPARATOR>({
+    separator,
+  });
 
   for (const [id, provider] of Object.entries(providers)) {
-    registry.registerProvider({ id, provider });
+    registry.registerProvider({ id, provider } as {
+      id: keyof PROVIDERS;
+      provider: PROVIDERS[keyof PROVIDERS];
+    });
   }
 
   return registry;
 }
 
-class DefaultProviderRegistry implements Provider {
-  private providers: Record<string, ProviderV1> = {};
+/**
+ * @deprecated Use `createProviderRegistry` instead.
+ */
+export const experimental_createProviderRegistry = createProviderRegistry;
 
-  registerProvider({
+class DefaultProviderRegistry<
+  PROVIDERS extends Record<string, ProviderV1>,
+  SEPARATOR extends string,
+> implements ProviderRegistryProvider<PROVIDERS, SEPARATOR>
+{
+  private providers: PROVIDERS = {} as PROVIDERS;
+  private separator: SEPARATOR;
+
+  constructor({ separator }: { separator: SEPARATOR }) {
+    this.separator = separator;
+  }
+
+  registerProvider<K extends keyof PROVIDERS>({
     id,
     provider,
   }: {
-    id: string;
-    provider: ProviderV1;
+    id: K;
+    provider: PROVIDERS[K];
   }): void {
     this.providers[id] = provider;
   }
 
   private getProvider(id: string): ProviderV1 {
-    const provider = this.providers[id];
+    const provider = this.providers[id as keyof PROVIDERS];
 
     if (provider == null) {
       throw new NoSuchProviderError({
@@ -49,7 +114,7 @@ class DefaultProviderRegistry implements Provider {
     id: string,
     modelType: 'languageModel' | 'textEmbeddingModel' | 'imageModel',
   ): [string, string] {
-    const index = id.indexOf(':');
+    const index = id.indexOf(this.separator);
 
     if (index === -1) {
       throw new NoSuchModelError({
@@ -57,14 +122,16 @@ class DefaultProviderRegistry implements Provider {
         modelType,
         message:
           `Invalid ${modelType} id for registry: ${id} ` +
-          `(must be in the format "providerId:modelId")`,
+          `(must be in the format "providerId${this.separator}modelId")`,
       });
     }
 
     return [id.slice(0, index), id.slice(index + 1)];
   }
 
-  languageModel(id: string): LanguageModel {
+  languageModel<KEY extends keyof PROVIDERS>(
+    id: `${KEY & string}${SEPARATOR}${string}`,
+  ): LanguageModel {
     const [providerId, modelId] = this.splitId(id, 'languageModel');
     const model = this.getProvider(providerId).languageModel?.(modelId);
 
@@ -75,7 +142,9 @@ class DefaultProviderRegistry implements Provider {
     return model;
   }
 
-  textEmbeddingModel(id: string): EmbeddingModel<string> {
+  textEmbeddingModel<KEY extends keyof PROVIDERS>(
+    id: `${KEY & string}${SEPARATOR}${string}`,
+  ): EmbeddingModel<string> {
     const [providerId, modelId] = this.splitId(id, 'textEmbeddingModel');
     const provider = this.getProvider(providerId);
 
@@ -91,7 +160,9 @@ class DefaultProviderRegistry implements Provider {
     return model;
   }
 
-  imageModel(id: string): ImageModel {
+  imageModel<KEY extends keyof PROVIDERS>(
+    id: `${KEY & string}${SEPARATOR}${string}`,
+  ): ImageModel {
     const [providerId, modelId] = this.splitId(id, 'imageModel');
     const provider = this.getProvider(providerId);
 
@@ -102,12 +173,5 @@ class DefaultProviderRegistry implements Provider {
     }
 
     return model;
-  }
-
-  /**
-   * @deprecated Use `textEmbeddingModel` instead.
-   */
-  textEmbedding(id: string): EmbeddingModel<string> {
-    return this.textEmbeddingModel(id);
   }
 }