+++ packages/google-vertex/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ai-sdk/google-vertex
 
+## 2.2.5
+
+### Patch Changes
+
+- 9507f7e: fix (provider/google-vertex): pass through auth options for vertex provider
+
+## 2.2.4
+
+### Patch Changes
+
+- Updated dependencies [aeaa92b]
+  - @ai-sdk/anthropic@1.2.2
+
 ## 2.2.3
 
 ### Patch Changes